small refactor, support -- in submit
If submit includes a file in the current directory, this file would be read as the potential script. If you want ot call this script (as opposed to submitting it), you can add -- before the script name.

ex:

batchq submit -- ./foo.sh bar

will now submit "./foo.sh bar" as the command to run.